nextjs restructure
