+++ src/pages/integration-builder/builder/frameworks/nextjs/stepContent/buildingApp.mdx
